feat(flow): add eval() interface generates hanging pod
